moved the dump loader from the backend to shared/
